Kokkos library update to v2.03.05

+++ lib/kokkos/core/src/KokkosExp_MDRangePolicy.hpp
@@ -236,13 +236,13 @@ struct MDRangePolicy
     MDRangePolicy( lower_tmp, upper_tmp, tile_tmp );
 
 #else
-    if(m_lower.size()!=rank || m_upper.size() != rank)
+    if(static_cast<int>(m_lower.size()) != rank || static_cast<int>(m_upper.size()) != rank)
       Kokkos::abort("MDRangePolicy: Constructor initializer lists have wrong size");
 
     for ( auto i = 0; i < rank; ++i ) {
       m_lower[i] = static_cast<array_index_type>(lower.begin()[i]);
       m_upper[i] = static_cast<array_index_type>(upper.begin()[i]);
-      if(tile.size()==rank)
+      if(static_cast<int>(tile.size())==rank)
         m_tile[i] = static_cast<array_index_type>(tile.begin()[i]);
       else
         m_tile[i] = 0;
